Fix review items.
