+++ nifi-api/src/main/java/org/apache/nifi/expression/AttributeExpression.java
@@ -75,8 +75,7 @@ public interface AttributeExpression {
      */
     ResultType getResultType();
 
-    public static enum ResultType {
-
+    enum ResultType {
         STRING, BOOLEAN, WHOLE_NUMBER, DATE, INSTANT, DECIMAL, NUMBER;
     }
 }
